# packages/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent Platform API",
    "version": "1.0.0",
    "description": "API for Agent Platform - a chat application with session management and real-time events"
  },
  "paths": {
...
    "/sessions": {
      "get": {
        "operationId": "sessions.list",
        "summary": "获取会话列表",
        "description": "获取所有会话的列表",
        "tags": [
          "Sessions"
        ],
        "responses": {
          "200": {
            "description": "会话列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sessions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "description": "会话 ID"
                          },
                          "name": {
                            "type": "string",
                            "description": "会话名称"
                          },
                          "createdAt": {
                            "type": "string",
                            "description": "创建时间"
                          },
                          "updatedAt": {
                            "type": "string",
                            "description": "更新时间"
                          },
                          "messageCount": {
                            "type": "number",
                            "description": "消息数量"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "createdAt",
                          "updatedAt",
                          "messageCount"
                        ]
                      },
                      "description": "会话列表"
                    }
                  },
                  "required": [
                    "sessions"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "sessions.create",
        "summary": "创建新会话",
        "description": "创建一个新的聊天会话",
        "tags": [
          "Sessions"
        ],
        "responses": {
          "201": {
            "description": "创建的会话",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "session": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "description": "会话 ID"
                        },
                        "name": {
                          "type": "string",
                          "description": "会话名称"
                        },
                        "createdAt": {
                          "type": "string",
                          "description": "创建时间"
                        },
                        "updatedAt": {
                          "type": "string",
                          "description": "更新时间"
                        },
                        "messageCount": {
                          "type": "number",
                          "description": "消息数量"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "createdAt",
                        "updatedAt",
                        "messageCount"
                      ],
                      "description": "创建的会话"
                    }
                  },
                  "required": [
                    "session"
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "description": "会话名称",
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/sessions/{id}/fork": {
      "post": {
        "operationId": "sessions.fork",
        "summary": "Fork 会话",
        "description": "从指定消息的某个位置创建一个新会话，保留该位置之前的所有消息",
        "tags": [
          "Sessions"
        ],
        "responses": {
          "201": {
            "description": "Fork 成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "session": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "description": "会话 ID"
                        },
                        "name": {
                          "type": "string",
                          "description": "会话名称"
                        },
                        "createdAt": {
                          "type": "string",
                          "description": "创建时间"
                        },
                        "updatedAt": {
                          "type": "string",
                          "description": "更新时间"
                        },
                        "messageCount": {
                          "type": "number",
                          "description": "消息数量"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "createdAt",
                        "updatedAt",
                        "messageCount"
                      ],
                      "description": "新创建的会话"
                    }
                  },
                  "required": [
                    "session"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "会话或消息不存在",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "错误信息"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "messageId": {
                    "type": "string",
                    "description": "消息 ID"
                  },
                  "name": {
                    "description": "新会话名称",
                    "type": "string"
                  }
                },
                "required": [
                  "messageId"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ]
      }
    },
    "/sessions/{id}/resume": {
      "post": {
        "operationId": "sessions.resume",
        "summary": "Resume 会话",
        "description": "在指定消息位置之后恢复对话，截断该消息之后的所有消息",
        "tags": [
          "Sessions"
        ],
        "responses": {
          "200": {
            "description": "Resume 成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "操作是否成功"
                    }
                  },
                  "required": [
                    "success"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "会话或消息不存在",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "错误信息"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "messageId": {
                    "type": "string",
                    "description": "消息 ID"
                  }
                },
                "required": [
                  "messageId"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ]
      }
    },
    "/sessions/{id}/messages": {
      "post": {
        "operationId": "sessions.messages.add",
        "summary": "添加消息到会话",
        "description": "向指定会话添加一条消息",
        "tags": [
          "Messages"
        ],
        "responses": {
          "200": {
            "description": "添加成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "操作是否成功"
                    }
                  },
                  "required": [
                    "success"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "会话不存在",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "错误信息"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "role": {
                    "type": "string",
                    "enum": [
                      "user",
                      "assistant",
                      "system"
                    ],
                    "description": "消息角色"
                  },
                  "content": {
                    "type": "string",
                    "description": "消息内容"
                  },
                  "timestamp": {
                    "description": "消息时间",
                    "type": "string"
                  }
                },
                "required": [
                  "role",
                  "content"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ]
      }
    },
...
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ]
}